pyproject.toml
[project]
name = "shopsmart-support-agent"
version = "1.0.0"
description = "AI Customer Support Agent using MCP + OpenRouter"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.6.0",
    "openai>=2.0.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "langfuse>=4.5.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "anyio>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src", "app"]
omit = [
    # Require live MCP subprocess + LLM API — covered by integration tests
    "src/agent/runner.py",
    # Eval framework requires live agent + judge LLM
    "src/eval/evaluator.py",
    "src/eval/run_eval.py",
    "src/eval/test_cases.py",
    # Module entry point, trivial one-liner
    "src/mcp_server/__main__.py",
]

[tool.coverage.report]
show_missing = true
fail_under = 80

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
